Move calculate_diff to utils
