+++ websocket/crates/domain/src/lib.rs
@@ -1,7 +1,9 @@
 pub mod project;
 pub mod repository;
 pub mod types;
-pub mod utils;
 pub use types::projection;
 pub use types::snapshot;
 pub use types::user;
+mod utils;
+pub use utils::calculate_diff;
+pub use utils::generate_id;
